requirements.txt
Django>=4.1,<5.0
django-view-breadcrumbs
djangorestframework>=3.14.0
djangorestframework-camel-case>=1.3.0
uWSGI
Pillow
